+++ app/src/main/kotlin/com/revolgenx/anilib/list/data/service/MediaListService.kt
@@ -3,6 +3,7 @@ package com.revolgenx.anilib.list.data.service
 import com.revolgenx.anilib.common.data.model.PageModel
 import com.revolgenx.anilib.list.data.field.MediaListCollectionField
 import com.revolgenx.anilib.list.data.field.MediaListCollectionIdField
+import com.revolgenx.anilib.list.data.field.MediaListCompareField
 import com.revolgenx.anilib.list.data.field.MediaListField
 import com.revolgenx.anilib.list.ui.model.MediaListCollectionModel
 import com.revolgenx.anilib.list.ui.model.MediaListModel
@@ -12,4 +13,5 @@ interface MediaListService {
     fun getMediaListCollectionsIds(field: MediaListCollectionIdField): Flow<List<Int>>
     fun getMediaListCollection(field: MediaListCollectionField): Flow<MediaListCollectionModel>
     fun getMediaList(field: MediaListField): Flow<PageModel<MediaListModel>>
+    fun getMediaListCompare(field: MediaListCompareField): Flow<PageModel<MediaListModel>>
 }