+++ app/src/main/kotlin/com/revolgenx/anilib/list/ui/viewmodel/MediaListViewModel.kt
@@ -256,4 +256,6 @@ abstract class MediaListViewModel(
         }
     }
 
+    fun getRandomMedia() = mediaListCollection.randomOrNull()
+
 }