+++ app/src/main/kotlin/com/revolgenx/anilib/setting/ui/screen/WidgetSettingsScreen.kt
@@ -18,6 +18,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Slider
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableFloatStateOf
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
@@ -31,14 +32,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.glance.appwidget.updateAll
 import anilib.i18n.R
-import com.maxkeppeker.sheets.core.models.base.rememberUseCaseState
-import com.maxkeppeler.sheets.color.ColorDialog
-import com.maxkeppeler.sheets.color.models.ColorConfig
-import com.maxkeppeler.sheets.color.models.ColorSelection
-import com.maxkeppeler.sheets.color.models.MultipleColors
-import com.maxkeppeler.sheets.color.models.SingleColor
 import com.revolgenx.anilib.common.ext.horizontalBottomWindowInsets
 import com.revolgenx.anilib.common.ext.localContext
+import com.revolgenx.anilib.common.ui.color.ColorPickerDialog
 import com.revolgenx.anilib.common.ui.component.menu.SortMenuItem
 import com.revolgenx.anilib.common.ui.component.menu.SortOrder
 import com.revolgenx.anilib.common.ui.component.menu.SortSelectMenu
@@ -227,7 +223,6 @@ private fun WidgetSettingsScreenContent() {
 
 
 @Composable
-@OptIn(ExperimentalMaterial3Api::class)
 private fun ColorGroupContent(
     color: Int,
     title: String,
@@ -236,24 +231,24 @@ private fun ColorGroupContent(
     colorsInt: List<Int> = listOf(),
     onColorSelected: OnClickWithValue<Int>
 ) {
-    val state = rememberUseCaseState()
-    ColorDialog(
-        state = state,
-        selection = ColorSelection(
-            selectedColor = SingleColor(colorInt = color),
-            onSelectColor = { selectedColor ->
-                onColorSelected(selectedColor)
-            }
-        ),
-        config = ColorConfig(
-            templateColors = MultipleColors.ColorsInt(colorsInt)
-        ))
+
+    val openColorPickerDialog = remember {
+        mutableStateOf(false)
+    }
+
+    ColorPickerDialog(
+        openDialog = openColorPickerDialog,
+        title = title,
+        selectedColor = color,
+        colorsInt = colorsInt.toTypedArray(),
+        onColorSelected = onColorSelected
+    )
     Row(
         modifier = Modifier
             .fillMaxWidth()
             .alpha(if (enabled) 1f else 0.5f)
             .clickable(enabled = enabled) {
-                state.show()
+                openColorPickerDialog.value = true
             },
         verticalAlignment = Alignment.CenterVertically
     ) {